+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Nothing
 ## Other
 
 * Avoid creation of unused archetypes by splitting the archetype graph out of the actual archetypes (#113)
+* Use slice instead of fixed-size array for type lookup in component registry (#113)
 
 ## [[v0.4.1]](https://github.com/mlange-42/arche/compare/v0.4.0...v0.4.1)